backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pyyaml
jsonschema
prometheus-client
python-dotenv
pytest
pytest-asyncio
